`AttitudeDynamics`のモジュールを作った #6
